Move configs to packages
