+++ .env.development
@@ -1,4 +1,4 @@
 # This is a stub.
 # It is needed as a data sample for TypeScript & Typechecking.
-# The real value of the variable is set in /bin/watch.js and depend on /config/renderer.vite.js
+# The real value of the variable is set in /bin/watch.js and depend on /config/vite.config.ts
 VITE_DEV_SERVER_URL=http://localhost:3000/
