+++ config/electron-vendors.js
@@ -1,4 +1,7 @@
-module.exports = {
-  chrome: 89,
-  node: 14,
-};
+export const chrome = 89;
+export const node = 14;
+
+  // = {
+  // chrome: 89,
+  // node: 14,
+// };
