+++ packages/main/src/index.ts
