+++ packages/preload/tsconfig.json
@@ -10,7 +10,7 @@
     "baseUrl": ".",
     "paths": {
       "/@/*": [
-        "./*"
+        "./src/*"
       ]
     },
     "types": [
@@ -19,7 +19,7 @@
     ]
   },
   "files": [
-    "./index.ts"
+    "./src/index.ts"
   ],
   "exclude": [
     "**/node_modules",
