+++ packages/renderer/assets/logo.png
