+++ packages/renderer/index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
 <div id="app"></div>
-<script src="./index.ts" type="module"></script>
+<script src="./src/index.ts" type="module"></script>
 </body>
 </html>
