+++ packages/renderer/src/components/About.vue
@@ -16,7 +16,7 @@
 
 <script lang="ts">
 import {defineComponent} from 'vue';
-import {useElectron} from '/@/use/electron';
+import {useElectron} from '../use/electron';
 export default defineComponent({
   name: 'App',
   setup() {
