+++ packages/renderer/src/components/AppNavigation.vue
