+++ packages/renderer/src/index.ts
@@ -1,6 +1,6 @@
 import {createApp} from 'vue';
 import App from './App.vue';
-import router from '/@/router';
+import router from './router';
 
 createApp(App)
   .use(router)
