+++ packages/renderer/src/use/electron.ts
@@ -1,4 +1,4 @@
-import type {ExposedInMainWorld} from '../../preload';
+import type {ExposedInMainWorld} from '../../../preload/src';
 
 export function useElectron(): ExposedInMainWorld {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
